backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.41
pydantic==2.11.4